% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_groups.R
\name{age_groups}
\alias{age_groups}
\title{age_groups}
\usage{
age_groups(data, column)
}
\arguments{
\item{data}{dataframe}

\item{column}{age column (in years)}
}
\value{
column with age categories
}
\description{
age_groups
}
\examples{
df <- data.frame(age = sample(18:65, size = 5, replace = TRUE))
df |> age_groups(age)
}
