% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{auto_scale_to_fit}
\alias{auto_scale_to_fit}
\title{Calculate auto-scaled parameters to fit content}
\usage{
auto_scale_to_fit(
  height_info,
  current_category_height,
  current_spacing,
  n_boxes
)
}
\arguments{
\item{height_info}{Result from calculate_required_height}

\item{current_category_height}{Current category height}

\item{current_spacing}{Current box spacing}

\item{n_boxes}{Number of boxes}
}
\value{
List with adjusted parameters
}
\description{
Calculate auto-scaled parameters to fit content
}
\keyword{internal}
