% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{calculate_subticks}
\alias{calculate_subticks}
\title{Calculate number of subticks}
\usage{
calculate_subticks(ratio)
}
\arguments{
\item{ratio}{Tick ratio}
}
\value{
Number of subticks
}
\description{
Calculate number of subticks
}
\keyword{internal}
