% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{check_character}
\alias{check_character}
\title{Check if value is a character string}
\usage{
check_character(x, name = "value", allow_null = FALSE)
}
\arguments{
\item{x}{Value to check}

\item{name}{Name of parameter for error message}

\item{allow_null}{Allow NULL values}
}
\description{
Check if value is a character string
}
\keyword{internal}
