% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{check_character_vector}
\alias{check_character_vector}
\title{Check if value is a character vector}
\usage{
check_character_vector(x, name = "value", min_length = 1, allow_null = FALSE)
}
\arguments{
\item{x}{Value to check}

\item{name}{Name of parameter for error message}

\item{min_length}{Minimum required length}

\item{allow_null}{Allow NULL values}
}
\description{
Check if value is a character vector
}
\keyword{internal}
