% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{check_layout_fit}
\alias{check_layout_fit}
\title{Check if layout will fit and provide recommendations}
\usage{
check_layout_fit(data, split_axis_by_col, split_box_by_col, options)
}
\arguments{
\item{data}{Data frame being plotted}

\item{split_axis_by_col}{Column for axis splitting}

\item{split_box_by_col}{Column for box splitting}

\item{options}{page_options object}
}
\value{
List with fit status and recommendations
}
\description{
Check if layout will fit and provide recommendations
}
\keyword{internal}
