% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{check_logical}
\alias{check_logical}
\title{Check if value is a logical}
\usage{
check_logical(x, name = "value")
}
\arguments{
\item{x}{Value to check}

\item{name}{Name of parameter for error message}
}
\description{
Check if value is a logical
}
\keyword{internal}
