% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{check_non_negative_numeric}
\alias{check_non_negative_numeric}
\title{Check if value is a non-negative numeric}
\usage{
check_non_negative_numeric(x, name = "value")
}
\arguments{
\item{x}{Value to check}

\item{name}{Name of parameter for error message}
}
\description{
Check if value is a non-negative numeric
}
\keyword{internal}
