% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{check_numeric_vector}
\alias{check_numeric_vector}
\title{Check if value is a numeric vector}
\usage{
check_numeric_vector(x, name = "value", min_length = 1)
}
\arguments{
\item{x}{Value to check}

\item{name}{Name of parameter for error message}

\item{min_length}{Minimum required length}
}
\description{
Check if value is a numeric vector
}
\keyword{internal}
