% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{check_page_options}
\alias{check_page_options}
\title{Check if object is a page_options instance}
\usage{
check_page_options(x, name = "options")
}
\arguments{
\item{x}{Object to check}

\item{name}{Name of parameter for error message}
}
\description{
Check if object is a page_options instance
}
\keyword{internal}
