% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{check_proportion}
\alias{check_proportion}
\title{Check if value is in range \link{0, 1}}
\usage{
check_proportion(x, name = "value")
}
\arguments{
\item{x}{Value to check}

\item{name}{Name of parameter for error message}
}
\description{
Check if value is in range \link{0, 1}
}
\keyword{internal}
