% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square.R
\name{chi_square}
\alias{chi_square}
\title{chi_square}
\usage{
chi_square(df, type)
}
\arguments{
\item{df}{dataframe}

\item{type}{condition column}
}
\value{
chi-squared htest class
}
\description{
chi_square
}
\examples{
\dontrun{
chi_square(df, 'biased')
}
}
