% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square_total.R
\name{chi_square_total}
\alias{chi_square_total}
\title{chi_square_total}
\usage{
chi_square_total(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
chi-squared htest class
}
\description{
This does chi square for all conditions combined.
}
\examples{
\dontrun{
chi_square_total(df)
}
}
