% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_label.R
\name{cond_label}
\alias{cond_label}
\title{cond_label}
\usage{
cond_label(data, column_name)
}
\arguments{
\item{data}{dataframe}

\item{column_name}{condition label, e.g. 'FD1', 'FL1' or 'B1'}
}
\value{
dataframe with column named `condition` where `FD1` becomes `fair_dark`, `FL1` becomes `fair_light`, and `B1` becomes `biased`.
}
\description{
cond_label
}
\examples{
\dontrun{
df |> cond_label(column_name = condition)
}
}
