% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_prefix.R
\docType{data}
\name{cond_prefix}
\alias{cond_prefix}
\title{Conditions Prefix}
\format{
A vector with 6 values:
\describe{
  \item{FD1_}{Fair-Dark Condition}
  \item{FD2_}{Fair-Dark Counterbalanced}
  \item{FL1_}{Fair-Light Condition}
  \item{FL2_}{Fair-Light Counterbalanced}
  \item{B1_}{Biased Condition}
  \item{B2_}{Biased Counterbalanced}

}
}
\source{
\url{data_URL}
}
\usage{
cond_prefix
}
\description{
Description of the dataset.
}
\keyword{datasets}
