% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_prob.R
\name{expected_prob}
\alias{expected_prob}
\title{expected_prob}
\usage{
expected_prob(data, condition)
}
\arguments{
\item{data}{dataframe}

\item{condition}{experiment condition}
}
\value{
rounded half of the number of instances. E.g. If a condition was in 501 trials, then 250 will be returned.
}
\description{
Given a condition in the experiment, find the number of times that condition was one of the trials.
}
\examples{
\dontrun{
expected_prob(df, 'FD1')
}
}
