% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{format_log_tick}
\alias{format_log_tick}
\title{Format logarithmic tick labels}
\usage{
format_log_tick(x, b, nice = TRUE)
}
\arguments{
\item{x}{Value to format}

\item{b}{Base of logarithm}

\item{nice}{Use nice formatting}
}
\value{
Formatted label
}
\description{
Format logarithmic tick labels
}
\keyword{internal}
