% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cols_between.R
\name{get_cols_between}
\alias{get_cols_between}
\title{get_cols_between}
\usage{
get_cols_between(df, col_start, col_end)
}
\arguments{
\item{df}{dataframe}

\item{col_start}{column to start from}

\item{col_end}{last column you want}
}
\value{
the names of the columns between the range
}
\description{
get_cols_between
}
\examples{
\dontrun{
get_cols_between(df = mtcars, col_start = mpg, col_end = hp)
}
}
