% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id_vals.R
\name{get_id_vals}
\alias{get_id_vals}
\title{get_id_vals}
\usage{
get_id_vals(data, id)
}
\arguments{
\item{data}{dataframe}

\item{id}{id vector to search render_id with}
}
\value{
dataframe of the render id
}
\description{
get_id_vals
}
\examples{
id_vals <- c(34, 95, 11)
df <- data.frame(render_id = c(34, 2, 35, 10, 11, 5))
get_id_vals(df, id_vals)
}
