% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_cond.R
\name{get_model_cond}
\alias{get_model_cond}
\title{get_model_cond}
\usage{
get_model_cond(data, column_name)
}
\arguments{
\item{data}{dataframe}

\item{column_name}{column name for the conditions}
}
\value{
dataframe with labels for the conditions, where condition == 'FD1' returns fair for cond1 and dark for cond2
}
\description{
get_model_cond
}
\examples{
df <- data.frame(val = c('FD1', 'FL1', 'FD1', 'B1'))
get_model_cond(df, val)
}
