% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt.R
\name{lrt}
\alias{lrt}
\title{lrt}
\usage{
lrt(chi_square_data)
}
\arguments{
\item{chi_square_data}{chi-squared htest class}
}
\value{
list of likelihood_ratio_statistic, degrees of freedom, p_value
}
\description{
lrt
}
\examples{
\dontrun{
x <- c(12, 5)
x_chi <- chisq.test(x)
lrt(x_chi)
}
}
