% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_drive.R
\name{parse_drive}
\alias{parse_drive}
\title{parse_drive}
\usage{
parse_drive(id)
}
\arguments{
\item{id}{document id from google drive}
}
\value{
document path to pass to a function that reads files
}
\description{
parse_drive
}
\examples{
parse_drive(id = '23492ff83')
}
