% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_condition.R
\name{participant_condition}
\alias{participant_condition}
\title{participant_condition}
\usage{
participant_condition(data, id_col, id_lst)
}
\arguments{
\item{data}{dataframe}

\item{id_col}{column of uuid}

\item{id_lst}{nested list with a list for 'fl' and 'fd'}
}
\value{
dataframe with a new column named `exp_cond` which specifies if the participant was in either the fair_light or fair_dark condition.
}
\description{
participant_condition
}
\examples{
\dontrun{
participant_condition(df, uuid, uuid_condition_lst)
}
}
