% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_constructs.R
\name{plot_dot}
\alias{plot_dot}
\title{Plot dots on a box}
\usage{
plot_dot(
  box,
  x,
  n1,
  n2 = 1,
  N2 = 1,
  pch = NULL,
  size = unit(1, "char"),
  col = "black"
)
}
\arguments{
\item{box}{The box object containing axis and position information.}

\item{x}{Numeric. The x-coordinate of the dot (on the data scale).}

\item{n1}{Integer. The row/category index within the box.}

\item{n2}{Integer. The sub-row position within the category. Default is 1.}

\item{N2}{Integer. The total number of sub-rows in the category. Default is 1.}

\item{pch}{Integer. The point character to use. Default is NULL (auto-select).
If N2 > 1, cycles through filled shapes (21-26).}

\item{size}{A unit object. The size of the point. Default is unit(1, 'char').}

\item{col}{Character. The color of the point. Default is 'black'.}
}
\value{
Invisible NULL. Draws on the current graphics device.
}
\description{
This function plots point markers on a box, typically used for point estimates
in forest plots.
}
\examples{
\dontrun{
# After creating a box object
plot_dot(box, x = 1.0, n1 = 1, col = 'purple')

# Multiple points in same category
plot_dot(box, x = 0.8, n1 = 1, n2 = 1, N2 = 2, col = 'blue')
plot_dot(box, x = 1.2, n1 = 1, n2 = 2, N2 = 2, col = 'red')
}

}
