% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_constructs.R
\name{plot_horizontal_seg}
\alias{plot_horizontal_seg}
\title{Plot a horizontal segment on a box plot}
\usage{
plot_horizontal_seg(
  box,
  x,
  n1,
  n2 = 1,
  N2 = 1,
  lty = 1,
  lwd = 3,
  col = "black"
)
}
\arguments{
\item{box}{The box object containing axis and position information.}

\item{x}{Numeric vector of length 2. The x-coordinates (start and end) of the segment.}

\item{n1}{Integer. The row/category index within the box.}

\item{n2}{Integer. The sub-row position within the category. Default is 1.}

\item{N2}{Integer. The total number of sub-rows in the category. Default is 1.}

\item{lty}{Integer. The line type. Default is 1 (solid).}

\item{lwd}{Numeric. The line width. Default is 3.}

\item{col}{Character. The line color. Default is 'black'.}
}
\value{
Invisible NULL. Draws on the current graphics device.
}
\description{
This function plots a horizontal segment (line) on a box plot, typically used
for confidence intervals in forest plots.
}
\examples{
\dontrun{
# After creating a box object
plot_horizontal_seg(box, x = c(0.5, 1.5), n1 = 1, col = 'blue')
}

}
