% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_constructs.R
\name{plot_long_rectangle}
\alias{plot_long_rectangle}
\title{Plot a long rectangle on a box plot}
\usage{
plot_long_rectangle(
  box,
  x,
  n1,
  n2 = 1,
  N2 = 1,
  height = 0.015,
  lty = 1,
  lwd = 3,
  col = "black"
)
}
\arguments{
\item{box}{The box object containing axis and position information.}

\item{x}{Numeric vector of length 2. The x-coordinates (start and end) of the rectangle.}

\item{n1}{Integer. The row/category index within the box.}

\item{n2}{Integer. The sub-row position within the category. Default is 1.}

\item{N2}{Integer. The total number of sub-rows in the category. Default is 1.}

\item{height}{Numeric. The height of the rectangle. Default is 0.015.}

\item{lty}{Integer. The line type for the border. Default is 1 (solid).}

\item{lwd}{Numeric. The line width for the border. Default is 3.}

\item{col}{Character. The color of the rectangle. Default is 'black'.}
}
\value{
Invisible NULL. Draws on the current graphics device.
}
\description{
This function plots a horizontal rectangle on a box plot, used as an alternative
to line segments for representing ranges.
}
\examples{
\dontrun{
# After creating a box object
plot_long_rectangle(box, x = c(0.5, 1.5), n1 = 1, height = 0.02, col = 'red')
}

}
