% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferred_model.R
\name{preferred_model}
\alias{preferred_model}
\title{preferred_model}
\usage{
preferred_model(data, value, type)
}
\arguments{
\item{data}{dataframe}

\item{value}{value for question 1}

\item{type}{condition}
}
\value{
dataframe of the preferred model from question 1
}
\description{
preferred_model
}
\examples{
df <- data.frame(
  quest2_val = sample(1:100, 10),
  condition = sample(c('fair_dark', 'fair_light', 'biased'), 10, replace = TRUE)
  )
preferred_model(df, quest2_val, condition)
}
