% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_questions.R
\name{rename_questions}
\alias{rename_questions}
\title{rename_questions}
\usage{
rename_questions(x, cond)
}
\arguments{
\item{x}{dataframe}

\item{cond}{condition}
}
\value{
Dataframe with renamed columns relating to the questions. For example, `Q1_1...25` = `FD1_1_1`, `Q1_2...26` = `FD1_1_2`
}
\description{
rename_questions
}
\examples{
\dontrun{
rename_questions(df, 'FD1')
}
}
