% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_survey_questions.R
\name{rename_survey_questions}
\alias{rename_survey_questions}
\title{rename_survey_questions}
\usage{
rename_survey_questions(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe with the survey questions renamed
}
\description{
rename_survey_questions
}
\examples{
\dontrun{
rename_survey_questions(df)
}
}
