% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R, R/plot_constructs.R,
%   R/utils_validation.R
\name{safe_get_color}
\alias{safe_get_color}
\title{Safely get color from palette (internal)}
\usage{
safe_get_color(palette, index, default = "black")

safe_get_color(palette, index, default = "black")

safe_get_color(palette, index, default = "black")
}
\arguments{
\item{palette}{Color palette}

\item{index}{Index to retrieve}

\item{default}{Default color if index out of bounds}
}
\description{
Safely get color from palette (internal)

Safely get color from palette

Safely get color from palette
}
\keyword{internal}
