% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_br.R
\name{safe_get_palette_color}
\alias{safe_get_palette_color}
\title{Safely get color from palette}
\usage{
safe_get_palette_color(palette, index)
}
\arguments{
\item{palette}{Color palette vector}

\item{index}{Index to retrieve}
}
\value{
Color string
}
\description{
Safely get color from palette
}
\keyword{internal}
