% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{safe_pretty}
\alias{safe_pretty}
\title{Safe pretty function for tick calculation}
\usage{
safe_pretty(x)
}
\arguments{
\item{x}{Numeric value}
}
\value{
Second element of pretty sequence or 1 if not available
}
\description{
Safe pretty function for tick calculation
}
\keyword{internal}
