% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skrrrahh.R
\name{skrrrahh}
\alias{skrrrahh}
\title{Play a short rap ad lib.}
\usage{
skrrrahh(sound = 26)
}
\arguments{
\item{sound}{character string or number specifying what sound to be played by 
either specifying one of the built in sounds, specifying the path to a wav 
file or specifying an url. The default is Gucci Mane saying "BRRR", sound 26.
If \code{sound} does not match any
of the sounds above, or is a valid path or url, a random sound will be
played. Currently \code{skrrrahh} can only handle http urls, https is not
supported.}
}
\description{
\code{skrrrahh} is exactly like \code{beep} from the beepr package 
(https://cran.r-project.org/web/packages/beepr/index.html). 
The only difference is in the automatically loaded sounds available. 
\code{skrrrahh} uses rap adlibs from [The Rap Board](http://therapboard.com/).
}
\details{
If \code{skrrrahh} is not able to play the sound, a warning is issued rather 
than an error. This is in order to not risk aborting or stopping the process 
that you wanted to get notified about.
}

