% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.brrr}
\alias{summary.brrr}
\title{Summary function for an object with class brrr}
\usage{
\method{summary}{brrr}(obj, round = 3)
}
\arguments{
\item{obj}{Object with class brrr}

\item{round}{Significant figures for the output}
}
\description{
Summary function for an object with class brrr
}
