% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{add_benefit_arrows}
\alias{add_benefit_arrows}
\title{Add benefit arrows to the header}
\usage{
add_benefit_arrows(
  obj,
  neutral_relative_x,
  direction = "up",
  labels = c("Favors LY", "Favors Placebo"),
  col = "#043099"
)
}
\arguments{
\item{obj}{An object containing the plot parameters and options.}

\item{neutral_relative_x}{The relative x-coordinate of the neutral position of the arrow.}

\item{direction}{The direction of the arrows. Default is 'up'.}

\item{labels}{The labels for the arrows. Default is c('Favors LY', 'Favors Placebo').}

\item{col}{The color of the arrows. Default is '#043099'.}
}
\value{
None
}
\description{
This function adds benefit arrows to the header of a plot. The arrows indicate the direction
of benefit, either favoring the last year (LY) or favoring the placebo. The arrows are drawn
using the grid package.
}
\examples{
\dontrun{
# Create a plot object
obj <- create_header(c(1, 2, 3))

# Add benefit arrows to the header
add_benefit_arrows(obj, neutral_relative_x = 0.5, direction = 'up',
                   labels = c('Favors LY', 'Favors Placebo'), col = '#043099')
}

}
