% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{add_legend}
\alias{add_legend}
\title{Add a legend to a plot}
\usage{
add_legend(
  legend_items,
  xpos,
  ypos,
  width,
  height,
  n_rows = NULL,
  n_cols = NULL,
  byrow = TRUE,
  label = "Legend",
  options = page_options$new()
)
}
\arguments{
\item{legend_items}{A list of legend items, where each item is a list with the following properties:
\describe{
\item{label}{The label for the legend item}
\item{type}{The type of the legend item (e.g., "line")}
\item{pch}{The point character for the legend item}
\item{col}{The color of the legend item}
\item{lty}{The line type of the legend item}
\item{lwd}{The line width of the legend item}
}}

\item{xpos}{The x position of the legend box}

\item{ypos}{The y position of the legend box}

\item{width}{The width of the legend box}

\item{height}{The height of the legend box}

\item{n_rows}{The number of rows in the legend box (optional)}

\item{n_cols}{The number of columns in the legend box (optional)}

\item{byrow}{A logical value indicating whether the legend items should be arranged by row (TRUE) or by column (FALSE)}

\item{label}{The label for the legend box (optional). Default is 'Legend'. When provided, the label is displayed at the top of the legend box.}

\item{options}{An object of class "page_options" containing additional options for the legend}
}
\value{
None
}
\description{
This function adds a legend to a plot based on the provided legend items.
}
