% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brrr.R
\name{brrr}
\alias{brrr}
\title{Title}
\usage{
brrr(
  x,
  z = NULL,
  y,
  t,
  maxit = 150,
  start = NULL,
  param = "Richardson",
  method = "Mle"
)
}
\arguments{
\item{x}{matrix of explanatory variables for log-relative risk}

\item{z}{matrix of explanatory variables for the nuisance parameter}

\item{y}{binary response variable}

\item{t}{binary risk factor}

\item{maxit}{maximum number of iteration for the estimation algorithm}

\item{start}{initial guess for the parameter. If null vector of zeros}

\item{param}{nuisance parameter. Possible values are "Richardson" or "Alternative"}

\item{method}{estimation method. Possible values are "Mle", "MeanBr" and "MedianBr"}
}
\value{

}
\description{
Title
}
