% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{create_header}
\alias{create_header}
\title{Create a header for a page}
\usage{
create_header(breaks_widths, labels = NULL, options = page_options$new())
}
\arguments{
\item{breaks_widths}{A numeric vector specifying the widths of the breaks relative to its total width.}

\item{labels}{A character vector specifying the labels for the breaks. Default is NULL.
If provided, the length of the labels must be the same as the length of breaks_widths.}

\item{options}{An instance of the page_options class specifying the options for the page.
Default is a new instance of the page_options class.}
}
\value{
The created header.
}
\description{
This function creates a header for a page, using the specified breaks widths,
labels, and options.
}
\examples{
\dontrun{
grid.newpage()
create_header(c(1, 2, 3), c("A", "B", "C"))
create_header(c(1, 2, 3), options = page_options$new())
}

}
