% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ops.R
\name{get_metadata}
\alias{get_metadata}
\title{Get Metadata for a Dataset}
\usage{
get_metadata(
  data,
  split_axis_by_col,
  axis_labels_col,
  split_box_by_col,
  vline_col
)
}
\arguments{
\item{data}{The dataset to calculate metadata for.}

\item{split_axis_by_col}{The column to split the data by on the y-axis.}

\item{axis_labels_col}{The column containing the axis labels.}

\item{split_box_by_col}{The column to split the data by in each box.}

\item{vline_col}{The column for vertical lines.}
}
\value{
A data frame containing the calculated metadata.
}
\description{
This function calculates metadata for a given dataset. The metadata includes
the number of unique categories, minimum and maximum values, whether the data
is reversed, and the number of subcategories. This information is crucial
for constructing the plot layout.
}
