% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_data}
\alias{mock_data}
\title{Mock data used in Benefit-Risk Visualization}
\format{
\subsection{\code{mock_data}}{

A data frame with 5 rows and 5 columns:
\describe{
\item{endpoint}{The endpoint of the study.}
\item{treatment}{The treatment arm of the study.}
\item{placebo}{The placebo arm of the study.}
\item{estimator}{Estimator name.}
\item{value}{The value of the estimator.}
\item{lower}{The lower bound of the estimator.}
\item{upper}{The upper bound of the estimator.}
}
}
}
\usage{
mock_data
}
\description{
Mock data used in Benefit-Risk Visualization
}
\keyword{datasets}
