% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_data}
\alias{mock_data}
\title{Mock Benefit-Risk Data}
\format{
A data frame with 15 rows and 12 columns:
\describe{
\item{endpoint}{The endpoint of the study.}
\item{treatment}{The treatment arm of the study.}
\item{placebo}{The placebo arm of the study.}
\item{estimator}{Estimator name.}
\item{axis_number}{The number of the axis for separating the axes.}
\item{value}{The value of the estimator.}
\item{lower}{The lower bound of the confidence interval.}
\item{upper}{The upper bound of the confidence interval.}
\item{reversed}{A logical indicating if the axis is reversed.}
\item{logscale}{A logical indicating if the axis is in log scale.}
\item{logbase}{The base of the logarithm.}
\item{col3}{A character column for the third column in the plot.}
}
}
\usage{
mock_data
}
\description{
A dataset containing mock data for benefit-risk assessment. This data is used
in examples and tests to demonstrate the functionality of the package.
}
\keyword{datasets}
