% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{page_options}
\alias{page_options}
\title{Page Options Class}
\description{
This R6 class defines and manages options for page layout and plotting.
It provides a centralized way to control various aspects of the plot,
such as margins, fonts, colors, and other graphical parameters.
}
\details{
The class is initialized with a set of default values for all options.
These options can be modified using the \code{set_option} method and
retrieved using the \code{get_option} method.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{PAGE_TOP_MARGIN}}{Top margin of the page.}

\item{\code{PAGE_BOTTOM_MARGIN}}{Bottom margin of the page.}

\item{\code{PAGE_LEFT_MARGIN}}{Left margin of the page.}

\item{\code{PAGE_RIGHT_MARGIN}}{Right margin of the page.}

\item{\code{HEADER_HEIGHT}}{Height of the header.}

\item{\code{HEADER_WIDTH}}{Width of the header.}

\item{\code{br_palette}}{Color palette for the plot.}

\item{\code{row.label.font.size}}{Font size for row labels.}

\item{\code{label.use.separation.line}}{Logical, whether to use a separation line for labels.}

\item{\code{label.font.size}}{Font size for labels.}

\item{\code{label.font.usecolors}}{Logical, whether to use colors for label fonts.}

\item{\code{header.label.font.size}}{Font size for header labels.}

\item{\code{header.label.color}}{Color for header labels.}

\item{\code{axis.label.font.size}}{Font size for axis labels.}

\item{\code{axis.ticks.font.size}}{Font size for axis ticks.}

\item{\code{axis.ticks.font.rotation}}{Rotation for axis ticks.}

\item{\code{axis.ticks.label.nice}}{Logical, whether to use nice labels for axis ticks.}

\item{\code{box.spacing}}{Spacing between boxes.}

\item{\code{box.category.height}}{Height of a single category in a box.}

\item{\code{box.fill.color}}{Fill color for boxes.}

\item{\code{legend.label.font.size}}{Font size for legend labels.}

\item{\code{legend.header.font.size}}{Font size for legend header.}

\item{\code{legend.fill.color}}{Fill color for the legend.}

\item{\code{forest.line.type}}{Line type for the forest plot.}

\item{\code{forest.line.width}}{Line width for the forest plot.}

\item{\code{forest.pch.shift}}{Shift for the pch in the forest plot.}

\item{\code{plot.width}}{Width of the plot in inches.}

\item{\code{plot.height}}{Height of the plot in inches.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-page_options-new}{\code{page_options$new()}}
\item \href{#method-page_options-get_option}{\code{page_options$get_option()}}
\item \href{#method-page_options-set_option}{\code{page_options$set_option()}}
\item \href{#method-page_options-clone}{\code{page_options$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-new"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-new}{}}}
\subsection{Method \code{new()}}{
Initialize the page options.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{A list of options to override the default values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-get_option"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-get_option}{}}}
\subsection{Method \code{get_option()}}{
Get an option value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$get_option(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the option to get.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The value of the option.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-set_option"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-set_option}{}}}
\subsection{Method \code{set_option()}}{
Set an option value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$set_option(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the option to set.}

\item{\code{value}}{The new value for the option.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-clone"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
