% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{page_options}
\alias{page_options}
\title{Define a class for page options}
\format{
An object of class \code{page_options}
}
\description{
This class represents the options for a page layout.
It includes properties for the page margins, header height, and color palette.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{PAGE_TOP_MARGIN}}{The top margin of the page.}

\item{\code{PAGE_BOTTOM_MARGIN}}{The bottom margin of the page.}

\item{\code{PAGE_LEFT_MARGIN}}{The left margin of the page.}

\item{\code{PAGE_RIGHT_MARGIN}}{The right margin of the page.}

\item{\code{HEADER_HEIGHT}}{The height of the header.}

\item{\code{HEADER_WIDTH}}{The width of the header.}

\item{\code{br_palette}}{The color palette for the page.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-page_options-new}{\code{page_options$new()}}
\item \href{#method-page_options-get_palette}{\code{page_options$get_palette()}}
\item \href{#method-page_options-set_palette}{\code{page_options$set_palette()}}
\item \href{#method-page_options-get_page_parameters}{\code{page_options$get_page_parameters()}}
\item \href{#method-page_options-get_page_parameter}{\code{page_options$get_page_parameter()}}
\item \href{#method-page_options-set_page_parameter}{\code{page_options$set_page_parameter()}}
\item \href{#method-page_options-clone}{\code{page_options$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-new"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$new(
  PAGE_TOP_MARGIN = 0.05,
  PAGE_BOTTOM_MARGIN = 0.05,
  PAGE_LEFT_MARGIN = 0.05,
  PAGE_RIGHT_MARGIN = 0.05,
  HEADER_HEIGHT = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PAGE_TOP_MARGIN}}{The top margin of the page.}

\item{\code{PAGE_BOTTOM_MARGIN}}{The bottom margin of the page.}

\item{\code{PAGE_LEFT_MARGIN}}{The left margin of the page.}

\item{\code{PAGE_RIGHT_MARGIN}}{The right margin of the page.}

\item{\code{HEADER_HEIGHT}}{The height of the header.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-get_palette"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-get_palette}{}}}
\subsection{Method \code{get_palette()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$get_palette()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-set_palette"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-set_palette}{}}}
\subsection{Method \code{set_palette()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$set_palette(palette)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-get_page_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-get_page_parameters}{}}}
\subsection{Method \code{get_page_parameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$get_page_parameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-get_page_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-get_page_parameter}{}}}
\subsection{Method \code{get_page_parameter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$get_page_parameter(parameter)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-set_page_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-set_page_parameter}{}}}
\subsection{Method \code{set_page_parameter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$set_page_parameter(parameter, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-clone"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
