% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{page_options}
\alias{page_options}
\title{Define a class for page options}
\format{
An object of class \code{page_options}
}
\description{
Define a class for page options

Define a class for page options
}
\details{
This class represents the options for a page layout.
It includes properties for the page margins, header height, and color palette.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{PAGE_TOP_MARGIN}}{The top margin of the page.}

\item{\code{PAGE_BOTTOM_MARGIN}}{The bottom margin of the page.}

\item{\code{PAGE_LEFT_MARGIN}}{The left margin of the page.}

\item{\code{PAGE_RIGHT_MARGIN}}{The right margin of the page.}

\item{\code{HEADER_HEIGHT}}{The height of the header.}

\item{\code{HEADER_WIDTH}}{The width of the header.}

\item{\code{row.label.font.size}}{The font size of the row labels.}

\item{\code{header.label.font.size}}{The font size of the header labels.}

\item{\code{axis.label.font.size}}{The font size of the axis labels.}

\item{\code{axis.ticks.font.size}}{The font size of the axis ticks.}

\item{\code{box.spacing}}{The spacing between boxes.}

\item{\code{box.fill.color}}{The color of the box fill.}

\item{\code{br_palette}}{The color palette for the page.}

\item{\code{legend.label.font.size}}{The font size of the legend labels.}

\item{\code{legend.header.font.size}}{The font size of the legend header.}

\item{\code{legend.fill.color}}{The color of the legend fill.}

\item{\code{forest.line.type}}{The line type for the forest plot.}

\item{\code{forest.line.width}}{The line width for the forest plot.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-page_options-new}{\code{page_options$new()}}
\item \href{#method-page_options-get_palette}{\code{page_options$get_palette()}}
\item \href{#method-page_options-set_palette}{\code{page_options$set_palette()}}
\item \href{#method-page_options-get_page_parameters}{\code{page_options$get_page_parameters()}}
\item \href{#method-page_options-get_page_parameter}{\code{page_options$get_page_parameter()}}
\item \href{#method-page_options-set_page_parameter}{\code{page_options$set_page_parameter()}}
\item \href{#method-page_options-get_label_font_size}{\code{page_options$get_label_font_size()}}
\item \href{#method-page_options-set_label_font_size}{\code{page_options$set_label_font_size()}}
\item \href{#method-page_options-get_header_font_size}{\code{page_options$get_header_font_size()}}
\item \href{#method-page_options-set_header_font_size}{\code{page_options$set_header_font_size()}}
\item \href{#method-page_options-get_axis_label_font_size}{\code{page_options$get_axis_label_font_size()}}
\item \href{#method-page_options-set_axis_label_fong_size}{\code{page_options$set_axis_label_fong_size()}}
\item \href{#method-page_options-get_axis_ticks_font_size}{\code{page_options$get_axis_ticks_font_size()}}
\item \href{#method-page_options-set_axis_ticks_font_size}{\code{page_options$set_axis_ticks_font_size()}}
\item \href{#method-page_options-get_box_spacing}{\code{page_options$get_box_spacing()}}
\item \href{#method-page_options-set_box_spacing}{\code{page_options$set_box_spacing()}}
\item \href{#method-page_options-clone}{\code{page_options$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-new"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-new}{}}}
\subsection{Method \code{new()}}{
Initialize the page options.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$new(
  PAGE_TOP_MARGIN = 0.05,
  PAGE_BOTTOM_MARGIN = 0.05,
  PAGE_LEFT_MARGIN = 0.05,
  PAGE_RIGHT_MARGIN = 0.05,
  HEADER_HEIGHT = convertY(unit(13, "mm"), unitTo = "npc", valueOnly = TRUE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PAGE_TOP_MARGIN}}{The top margin of the page.}

\item{\code{PAGE_BOTTOM_MARGIN}}{The bottom margin of the page.}

\item{\code{PAGE_LEFT_MARGIN}}{The left margin of the page.}

\item{\code{PAGE_RIGHT_MARGIN}}{The right margin of the page.}

\item{\code{HEADER_HEIGHT}}{The height of the header.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-get_palette"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-get_palette}{}}}
\subsection{Method \code{get_palette()}}{
Get the color palette for the page.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$get_palette()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The color palette for the page.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-set_palette"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-set_palette}{}}}
\subsection{Method \code{set_palette()}}{
Set the color palette for the page.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$set_palette(palette)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{palette}}{The color palette to set.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-get_page_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-get_page_parameters}{}}}
\subsection{Method \code{get_page_parameters()}}{
Get the page parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$get_page_parameters()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of the page parameters.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-get_page_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-get_page_parameter}{}}}
\subsection{Method \code{get_page_parameter()}}{
Get a specific page parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$get_page_parameter(parameter)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameter}}{The parameter to get.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The value of the parameter.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-set_page_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-set_page_parameter}{}}}
\subsection{Method \code{set_page_parameter()}}{
Set a specific page parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$set_page_parameter(parameter, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameter}}{The parameter to set.}

\item{\code{value}}{The value to set.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-get_label_font_size"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-get_label_font_size}{}}}
\subsection{Method \code{get_label_font_size()}}{
Get the font size of the row labels.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$get_label_font_size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The font size of the row labels.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-set_label_font_size"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-set_label_font_size}{}}}
\subsection{Method \code{set_label_font_size()}}{
Set the font size of the row labels.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$set_label_font_size(size)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{The font size to set.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-get_header_font_size"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-get_header_font_size}{}}}
\subsection{Method \code{get_header_font_size()}}{
Get the font size of the header labels.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$get_header_font_size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The font size of the header labels.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-set_header_font_size"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-set_header_font_size}{}}}
\subsection{Method \code{set_header_font_size()}}{
Set the font size of the header labels.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$set_header_font_size(size)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{The font size to set.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-get_axis_label_font_size"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-get_axis_label_font_size}{}}}
\subsection{Method \code{get_axis_label_font_size()}}{
Get the font size of the axis labels.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$get_axis_label_font_size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The font size of the axis labels.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-set_axis_label_fong_size"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-set_axis_label_fong_size}{}}}
\subsection{Method \code{set_axis_label_fong_size()}}{
Set the font size of the axis labels.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$set_axis_label_fong_size(size)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{The font size to set.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-get_axis_ticks_font_size"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-get_axis_ticks_font_size}{}}}
\subsection{Method \code{get_axis_ticks_font_size()}}{
Get the font size of the axis ticks.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$get_axis_ticks_font_size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The font size of the axis ticks.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-set_axis_ticks_font_size"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-set_axis_ticks_font_size}{}}}
\subsection{Method \code{set_axis_ticks_font_size()}}{
Set the font size of the axis ticks.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$set_axis_ticks_font_size(size)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{The font size to set.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-get_box_spacing"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-get_box_spacing}{}}}
\subsection{Method \code{get_box_spacing()}}{
Get the box spacing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$get_box_spacing()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The box spacing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-set_box_spacing"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-set_box_spacing}{}}}
\subsection{Method \code{set_box_spacing()}}{
Set the box spacing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$set_box_spacing(spacing)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{spacing}}{The spacing to set.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
None
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page_options-clone"></a>}}
\if{latex}{\out{\hypertarget{method-page_options-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{page_options$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
