% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{plot_axis}
\alias{plot_axis}
\title{Plot Axis}
\usage{
plot_axis(
  xlength,
  xpos,
  ypos,
  from,
  to,
  n_ticks,
  neutral_pos,
  label = NULL,
  logscale = FALSE,
  b = 10,
  show_axis = TRUE
)
}
\arguments{
\item{xlength}{The length of the axis line.}

\item{xpos}{The x-coordinate of the starting position of the axis line (0 to 1).}

\item{ypos}{The y-coordinate of the starting position of the axis line (0 to 1).}

\item{from}{The starting value of the axis range.}

\item{to}{The ending value of the axis range.}

\item{n_ticks}{The number of ticks on the axis.}

\item{neutral_pos}{The position of the neutral tick (0 on linear scale, 1 on logarithmic scale).}

\item{label}{The label for the axis.}

\item{logscale}{Logical value indicating whether to use logarithmic scale.}

\item{b}{The base for logarithmic scale.}

\item{show_axis}{Logical value indicating whether to show the axis.}
}
\value{
A list containing the following elements:
\itemize{
\item axis_function: A function that scales a value to the position on the axis.
\item from: The transformed starting value of the axis range.
\item to: The transformed ending value of the axis range.
\item length: The length of the axis line.
}
}
\description{
This function plots an axis with ticks and labels on a given plot.
}
\examples{
\dontrun{
# Example usage:
grid.newpage()
pp_lin_axis <- plot_axis(xlength = 0.8, xpos = 0.5, ypos = 0.5, 
         from = 0, to = 10, b = 10, n_ticks = 6, neutral_pos = 1)

# reversed axis
pp_rev_axis <- plot_axis(xlength = 0.8, xpos = 0.5, ypos = 0.4, 
         from = 10, to = 0, b = 10, n_ticks = 6, neutral_pos = 1)

# logaritmic axis
pp_log_axis <- plot_axis(xlength = 0.8, xpos = 0.5, ypos = 0.3, 
          from = 0.1, to = 1000, b = 10, n_ticks = 6, neutral_pos = 1, 
          logscale = TRUE, label = "from=0.1 to 1000")

# logarithmic reversed axis with label that has math expression
pp_log_rev_axis <- plot_axis(xlength = 0.8, xpos = 0.5, ypos = 0.2, 
          from = 1000, to = 0.1, b = 10, n_ticks = 6, neutral_pos = 1, 
          logscale = TRUE, label = 'from=1000 to 0.1', show_axis=TRUE)

# another logaritmic axis
pp_log_axis2 <- plot_axis(xlength = 0.8, xpos = 0.5, ypos = 0.1, 
          from = 1.12, to = 99, b = 10, n_ticks = 6, neutral_pos = 1, 
          logscale = TRUE, label = "from=1.12 to 99")
}

}
