% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{plot_box}
\alias{plot_box}
\title{Plot Box}
\usage{
plot_box(
  xpos,
  ypos,
  xlength,
  n_categories,
  single_category_height,
  neutral_pos,
  n_ticks,
  from,
  to,
  label = NULL,
  logscale = FALSE,
  b = 10,
  show_axis = TRUE,
  options = page_options$new()
)
}
\arguments{
\item{xpos}{The x-coordinate of the box's position.}

\item{ypos}{The y-coordinate of the box's position.}

\item{xlength}{The length of the box along the x-axis.}

\item{n_categories}{The number of categories in the box.}

\item{single_category_height}{The height of each category within the box.}

\item{neutral_pos}{The position of the neutral line within the box.}

\item{n_ticks}{The number of ticks on the axis.}

\item{from}{The starting value of the axis range.}

\item{to}{The ending value of the axis range.}

\item{label}{The label for the axis.}

\item{logscale}{Logical value indicating whether to use logarithmic scale.}

\item{b}{The base for logarithmic scale.}

\item{show_axis}{Logical value indicating whether to show the axis.}

\item{options}{The page options object.}
}
\value{
A list containing the following elements:
\itemize{
\item axis: The axis of the box.
\item name: The name of the box.
\item height: The height of the box.
\item y_pos: The y-positions of the categories in the box.
}
}
\description{
This function plots a box at the specified position with the specified dimensions.
}
\examples{
\dontrun{
grid.newpage()
box1 <- plot_box(0.5, 0.5, 1, 3, 0.05, 3, 5, 0, 1, expression(x[1]^2), FALSE)
box2 <- plot_box(0.4, 0.4, 1, 3, 0.05, 3, 5, 0, 1, expression(x[1]^2), FALSE)
box3 <- plot_box(0.6, 0.6, 1, 3, 0.01, 3, 5, 0, 1, NULL, FALSE)
box4 <- plot_box(0.5, 0.7, 1, 3, 0.05, 3, 5, 0, 1, expression(x[1]^2), FALSE, show_axis=FALSE)
}
}
