% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_br.R
\name{plot_br}
\alias{plot_br}
\title{Plot Benefit-Risk Data}
\usage{
plot_br(
  data,
  columns_specs,
  breaks_widths,
  split_axis_by_col,
  axis_labels_col,
  split_box_by_col,
  vline_col = NULL,
  neutral_pos = 3,
  num_ticks = 6,
  top_margin = NULL,
  userect = FALSE,
  arrow_labels = c("Favors\\nTreatment", "Favors\\nPlacebo"),
  value_collapse = rep(FALSE, length(columns_specs)),
  label_text_size = NULL,
  header_text_size = NULL,
  box_group = NULL,
  colors_by = NULL,
  options_br = page_options$new(),
  filename = NULL
)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{columns_specs}{A named character vector specifying the columns in the plot.}

\item{breaks_widths}{A numeric vector specifying the relative widths of the columns.}

\item{split_axis_by_col}{The column name to split the y-axis by.}

\item{axis_labels_col}{The column name for the axis labels.}

\item{split_box_by_col}{The column name to split the boxes by.}

\item{vline_col}{The column name for vertical lines (e.g., non-inferiority margin).}

\item{neutral_pos}{The position of the neutral line.}

\item{num_ticks}{The number of ticks on the axis.}

\item{top_margin}{The top margin of the plot.}

\item{userect}{Logical, whether to use rectangles for the forest plot.}

\item{arrow_labels}{A character vector for the benefit direction arrows.}

\item{value_collapse}{A logical vector indicating whether to collapse values in each column.}

\item{label_text_size}{A numeric vector to adjust the label text size.}

\item{header_text_size}{A numeric vector to adjust the header text size.}

\item{box_group}{An optional \code{plot_br} object to group with.}

\item{colors_by}{The column name to color the plot elements by.}

\item{options_br}{A \code{page_options} object with plot options.}

\item{filename}{An optional string specifying the filename to save the plot.}
}
\value{
A list containing the boxes in the plot, used options, and the last y-position.
}
\description{
This function creates a forest plot for benefit-risk assessment.
It visualizes the data using boxes and forest plots, and allows for extensive customization.
}
\examples{
\dontrun{
data(mock_data)
columns_specs <- c('Benefits'='endpoint', 
              'Treatment\n(N=100)'='treatment', 
              'Placebo\n(N=100)'='placebo',
              'Comparison\nHR or Odds Ratio\n(95\% CI)'='col3')
plot_br(mock_data, columns_specs, c(0.2, -0.1, 0.1, 0.2), 
        'estimator', 'endpoint', 'endpoint')
}

}
