% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_br.R
\name{plot_br}
\alias{plot_br}
\title{Plot BR}
\usage{
plot_br(
  data,
  columns_specs,
  breaks_widths,
  split_axis_by_col,
  axis_labels_col,
  split_box_by_col,
  neutral_pos = 3,
  num_ticks = 6,
  top_margin = NULL,
  userect = FALSE,
  arrow_labels = c("Favors\\nTreatment", "Favors\\nPlacebo"),
  value_collapse = rep(FALSE, length(columns_specs)),
  options_br = page_options$new()
)
}
\arguments{
\item{data}{The data frame containing the BR data.}

\item{columns_specs}{A character named vector specifying the column names in the data frame that correspond to the different columns in the plot.}

\item{breaks_widths}{A numeric vector specifying the widths of the breaks between the columns in the plot.}

\item{split_axis_by_col}{A character string specifying the column name in the data frame to split the plot axis by.}

\item{axis_labels_col}{A character string specifying the column name in the data frame to use as the axis labels.}

\item{split_box_by_col}{A character string specifying the column name in the data frame to split the boxes within each axis by.}

\item{neutral_pos}{A numeric value specifying the neutral position of the plot.}

\item{num_ticks}{An optional numeric value specifying the number of ticks on the plot axis.}

\item{top_margin}{An optional numeric value specifying the top margin of the plot.}

\item{userect}{A logical value specifying whether to use rectangles in the plot.}

\item{arrow_labels}{A character vector specifying the labels for the arrows in the plot.}

\item{value_collapse}{A logical vector specifying whether to collapse the values within each column/row.}

\item{options_br}{An optional page_options object specifying the options for the plot.}
}
\value{
A list containing the boxes in the plot and the last y-position of the plot.
}
\description{
This function plots BR (Bayesian Regression) data using a forest plot. It takes in various parameters to customize the plot.
}
\examples{
\dontrun{
#create header
breaks_widths <- c(0.2, -0.1, 0.1, 0.2)
columns_specs <- c('Benefits'='endpoint', 
              'Treatment\n(N=100)'='treatment', 
              'Placebo\n(N=100)'='placebo',
              'Comparison\nHR or Odds Ratio\n(95\% CI)'='col3')
value_collapse <- c(TRUE, FALSE, FALSE, FALSE)

split_axis_by_col <- 'estimator'
split_box_by_col <- 'endpoint'

# load data and add column
data(mock_data)
part1_data <- plot_br(mock_data, columns_specs, breaks_widths, 
split_axis_by_col, split_box_by_col, top_margin=NULL, value_collapse=value_collapse)
}

}
