% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_constructs.R
\name{plot_dot}
\alias{plot_dot}
\title{Plot dots}
\usage{
plot_dot(
  box,
  x,
  n1,
  n2 = 1,
  N2 = 1,
  pch = 21,
  size = unit(1, "char"),
  col = "black"
)
}
\arguments{
\item{box}{The box to plot the dots on.}

\item{x}{The x-coordinate of the dots.}

\item{n1}{The number of dots to plot.}

\item{n2}{The number of dots to plot (default is 1).}

\item{N2}{The total number of dots (default is 1).}

\item{pch}{The symbol to use for the dots (default is 19).}

\item{size}{The size of the dots (default is unit(1, 'char')).}

\item{col}{The color of the dots (default is 'black').}
}
\value{
None
}
\description{
This function plots dots on a box.
}
\examples{
\dontrun{
plot_dot(box = my_box, x = 0.5, n1 = 10, n2 = 2, N2 = 20, pch = 19, size = unit(1, 'char'), 
         col = 'black')
}
}
