% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_constructs.R
\name{plot_horizontal_seg}
\alias{plot_horizontal_seg}
\title{Plot a horizontal segment on a box plot}
\usage{
plot_horizontal_seg(
  box,
  x,
  n1,
  n2 = 1,
  N2 = 1,
  lty = 1,
  lwd = 3,
  col = "black"
)
}
\arguments{
\item{box}{The box plot object.}

\item{x}{The x-coordinate of the segment.}

\item{n1}{The y-coordinate of the segment start.}

\item{n2}{The y-coordinate of the segment end (default is 1).}

\item{N2}{The total number of segments (default is 1).}

\item{lty}{The line type (default is 1).}

\item{lwd}{The line width (default is 3).}

\item{col}{The line color (default is 'black').}
}
\value{
None
}
\description{
This function plots a horizontal segment on a box plot.
}
\examples{
\dontrun{
# Plot a horizontal segment on the box plot
plot_horizontal_seg(box, x = 1, n1 = 20, n2 = 30, N2 = 2, lty = 2, lwd = 2, col = 'red')
}

}
