% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_constructs.R
\name{plot_long_rectangle}
\alias{plot_long_rectangle}
\title{Plot a long rectangle on a given plot}
\usage{
plot_long_rectangle(
  box,
  x,
  n1,
  n2 = 1,
  N2 = 1,
  height = 0.015,
  lty = 1,
  lwd = 3,
  col = "black"
)
}
\arguments{
\item{box}{The plot box object containing axis information}

\item{x}{The x-coordinates of the rectangle}

\item{n1}{The index of the y-axis position to use for the rectangle}

\item{n2}{The position of the rectangle along the y-axis (default is 1)}

\item{N2}{The total number of positions along the y-axis (default is 1)}

\item{height}{The height of the rectangle (default is 0.015)}

\item{lty}{The line type of the rectangle (default is 1)}

\item{lwd}{The line width of the rectangle (default is 3)}

\item{col}{The color of the rectangle (default is 'black')}
}
\value{
None
}
\description{
This function plots a long rectangle on a given plot using the specified parameters.
}
\examples{
# Create a plot box object
\dontrun{
# Plot a long rectangle
plot_long_rectangle(box, x = c(0.2, 0.8), n1 = 1)
}
}
